refactor: list_page_frameのリファクタリング GT-2
