+++ lib/screens/curriculums/curriculums_screen.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 
-import '../widgets/header/header_widget.dart';
-import '../widgets/notification_bar/notification_bar_widget.dart';
-import '../widgets/side_menu/side_menu_widget.dart';
-import '../models/skill.dart';
-import '../models/table_items/curriculum_item.dart';
-import '../models/user.dart';
-import '../utils/constants.dart';
-import '../widgets/frame/list_page_frame.dart';
+import '../../widgets/header/header_widget.dart';
+import '../../widgets/notification_bar/notification_bar_widget.dart';
+import '../../widgets/side_menu/side_menu_widget.dart';
+import '../../models/skill.dart';
+import '../../models/table_items/curriculum_item.dart';
+import '../../models/user.dart';
+import '../../utils/constants.dart';
+import '../../widgets/frame/list_page_frame.dart';
 
 class CurriculumsScreen extends StatelessWidget {
   const CurriculumsScreen({Key? key}) : super(key: key);
